backend/requirements.txt
alembic==1.13.3
bcrypt==4.0.1
email-validator==2.2.0
fastapi==0.115.6
httpx==0.27.2
passlib[bcrypt]==1.7.4
psycopg[binary]==3.2.3
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
pytest==8.3.4
pytest-cov==6.0.0
sqlalchemy==2.0.36
uvicorn[standard]==0.34.0

